封装 PFC 逻辑
