+++ src/plugins/chat/bot.py
@@ -3,12 +3,10 @@ from typing import Dict, Any
 from ..moods.moods import MoodManager  # 导入情绪管理器
 from ...config.config import global_config
 from .message import MessageRecv
-from ..PFC.pfc_manager import PFCManager
-from .chat_stream import chat_manager
-from .only_message_process import MessageProcessor
 
 from src.common.logger_manager import get_logger
 from ..heartFC_chat.heartflow_processor import HeartFCProcessor
+from ..PFC.pfc_processor import PFCProcessor
 from ..utils.prompt_builder import Prompt, global_prompt_manager
 import traceback
 
@@ -25,10 +23,7 @@ class ChatBot:
         self._started = False
         self.mood_manager = MoodManager.get_instance()  # 获取情绪管理器单例
         self.heartflow_processor = HeartFCProcessor()  # 新增
-
-        # 创建初始化PFC管理器的任务，会在_ensure_started时执行
-        self.only_process_chat = MessageProcessor()
-        self.pfc_manager = PFCManager.get_instance()
+        self.pfc_processor = PFCProcessor()
 
     async def _ensure_started(self):
         """确保所有任务已启动"""
@@ -36,17 +31,6 @@ class ChatBot:
             logger.trace("确保ChatBot所有任务已启动")
 
             self._started = True
-
-    async def _create_pfc_chat(self, message: MessageRecv):
-        try:
-            chat_id = str(message.chat_stream.stream_id)
-            private_name = str(message.message_info.user_info.user_nickname)
-
-            if global_config.enable_pfc_chatting:
-                await self.pfc_manager.get_or_create_conversation(chat_id, private_name)
-
-        except Exception as e:
-            logger.error(f"创建PFC聊天失败: {e}")
 
     async def message_process(self, message_data: Dict[str, Any]) -> None:
         """处理转化后的统一格式消息
@@ -118,18 +102,7 @@ class ChatBot:
                         # 是否进入PFC
                         if global_config.enable_pfc_chatting:
                             logger.trace("进入PFC私聊处理流程")
-                            userinfo = message.message_info.user_info
-                            messageinfo = message.message_info
-                            # 创建聊天流
-                            logger.trace(f"为{userinfo.user_id}创建/获取聊天流")
-                            chat = await chat_manager.get_or_create_stream(
-                                platform=messageinfo.platform,
-                                user_info=userinfo,
-                                group_info=groupinfo,
-                            )
-                            message.update_chat_stream(chat)
-                            await self.only_process_chat.process_message(message)
-                            await self._create_pfc_chat(message)
+                            await self.pfc_processor.process_message(message_data)
                         # 禁止PFC，进入普通的心流消息处理逻辑
                         else:
                             logger.trace("进入普通心流私聊处理")
